First pass of an an2ln_cdb module for Heimdal.
This module accepts configuration via krb5.conf:

[realms]
	FOO.BAR = {
		auth_to_local = CDB:/path/to/file.db DEFAULT
	}

The CDB:/path/to/file.db directive specifies that authenticated
user names will be looked up in the CDB database /path/to/file.db
and if an entry exists then it will be used as the local user name.
If the name is empty then the aname maps to no local user even if
a following rule would have matched it.  The aname up with which
is looked is a full krb5 principal name including the realm and
all components.